requirements.txt
flask==3.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.38.1
requests==2.31.0
markdown2==2.4.10
google-cloud-core==2.4.1
google-cloud-storage==2.13.0
Pillow==10.1.0 